test파일 수정 2

+++ src/main/java/com/front/node/ModbusServerNode.java
@@ -79,7 +79,7 @@ public class ModbusServerNode extends InputOutputNode {
                 object.put("address", unitData.get("address"));
                 object.put("value", unitData.get("value"));
 
-                map.put((Integer) unitData.get("unitId"), object);
+                map.put(((Long)unitData.get("unitId")).intValue(), object);
                 log.info("mapdata : {}", map.toString());
             }
         }
@@ -139,8 +139,9 @@ public class ModbusServerNode extends InputOutputNode {
                     switch (functionCode) {
                         case 3:
                             int[] holdingregisters = new int[address + 100];
-                            int unitAddress = (int) map.get(unitId).get("address");
-                            int unitValue = (int) map.get(unitId).get("value");
+                            int unitAddress = ((Long)map.get(unitId).get("address")).intValue();
+                            int unitValue = ((Double)map.get(unitId).get("value")).intValue();
+                            
 
                             holdingregisters[unitAddress] = unitValue;
                             if (address + quantity < holdingregisters.length) {
